Handle double-click on nick list and middle-click on channel list.
Double-click on nick list will open a private chat window and
switch to that window.

Middle-click on channel list will close the associated chat window,
and if it is a channel, issue a PART command to leave the channel.

As part of this, make sure the events only correspond to the
mouse button involved (use the event's "which" property, having
well-defined historical values of 1/2/3 left/mid/right). This
means that the previous channel click action now only corresponds
to the left (primary) button.