Don't continue completion if cursor moves; don't start anew after a space.
This prevents autocompletion from kicking in on a previously typed word
followed by a space, as well as when the completion is accepted and the
user begins to type more text. However, if the cursor is moved to the
end of a word without an intervening space, completion will start over
from scratch (which is a bit more intuitive).

Fixes #97.