Added \u0001 as an acceptable mentioning suffix
When a nick is at the end of an action, the \u0001 character is directly
after the nick. In these cases, the mention notification regex would not
match.

This commit also includes tests for different combinations of where the
nick can be with respect to other text in the action.