Document identifying to NickServ using `/server`
What was done?
The help description for the server command was modified to indicate to the user that he/she could identify to NickServ by pasing extra parameters to the server command itself.

Why was it done?
It is currently not easy for a user to auto-identify to NickServ on CIRC's start-up. An 'auto_identify.js' script does exist, but currently has to be downloaded seperately (timeline states intention to package it with CIRC in the future) with no obvious clue to the user that he has to install the script to make it work.

Where was it done?
/bin/chat/user_command_handler.js:82