Use native chrome ssl sockets
Previously, we relied on a pure javscript ssl implementation called forge
for secure sockets. This had a couple of significant downsides:
1. Certificate Authentication via your trusted CA (or any CA) didn't work
2. Name verification wasn't enabled

As a result, secure sockets were not secure.

This solves the problem by using the (present since Chrome 38) secure
socket support. This gives you all of the TLS certificate authentication
goodness of chrome for free. As a plus, it's also a relatively simple
change, so I was able to delete most of the logic for ssl and just add a
separate callback for pre secure connect.